feat(testing): expose fp8 quantization helpers via flashinfer.testing.fp8
Move the four FP8 quantization helpers (per_token_cast_to_fp8,
per_block_cast_to_fp8, quantize_fp8, dequantize_fp8) plus the
internal _ceil_to_ue8m0 helper from flashinfer/testing/utils.py into a
new flashinfer/testing/fp8 submodule, and document them under
docs/api/testing.rst.

Eight in-tree call sites currently import these helpers via the
semi-private path 'from flashinfer.testing.utils import ...' (e.g.
benchmarks/bench_deepgemm_blackwell.py, benchmarks/routines/moe.py,
benchmarks/routines/moe_comm.py, tests/gemm/test_fp8_blockscale_gemm.py).
Promoting them under a submodule gives downstream FP8-related
contributions a stable, documented public surface without expanding
the top-level flashinfer.testing namespace. The bare names remain
re-exported from flashinfer.testing.utils so existing callers keep
working unchanged.

No behavior changes.
