+++ docs/api/testing.rst
@@ -49,3 +49,24 @@ GPU Benchmarking
     bench_gpu_time_with_cuda_event
     bench_gpu_time_with_cudagraph
     bench_gpu_time_with_cupti
+
+FP8 Quantization Helpers
+------------------------
+
+Reference quantize/dequantize helpers used by FP8 tests and benchmarks. Import
+from the ``flashinfer.testing.fp8`` submodule:
+
+.. code-block:: python
+
+    from flashinfer.testing.fp8 import quantize_fp8, dequantize_fp8
+    from flashinfer.testing.fp8 import per_token_cast_to_fp8, per_block_cast_to_fp8
+
+.. currentmodule:: flashinfer.testing.fp8
+
+.. autosummary::
+    :toctree: ../generated
+
+    per_token_cast_to_fp8
+    per_block_cast_to_fp8
+    quantize_fp8
+    dequantize_fp8
